+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 #### Bugfixes
 
+- fixed windows filepath issue. ([#178](https://github.com/peak/s5cmd/issues/178)) [@mrekim0](https://github.com/mrekim0)
+- windows tests pass.
+
+#### Bugfixes
+
 - Fixed a transfer issue where the object path contains particular regex metacharacters. ([#111](https://github.com/peak/s5cmd/pull/111)) [@brendan-matroid](https://github.com/brendan-matroid)
 - Correctly parse object paths that contain whitespaces in run-mode. ([#111](https://github.com/peak/s5cmd/pull/111)) [@brendan-matroid](https://github.com/brendan-matroid)